TIP-1094: use a simple subscriber to archive import/export logs
