+++ src/Akeneo/Tool/Bundle/ConnectorBundle/Resources/config/archiving.yml
@@ -71,9 +71,9 @@ services:
         tags:
             - { name: pim_connector.archiver }
 
-    pim_connector.archiver.log_archiver:
-        class: 'Akeneo\Tool\Component\Connector\Archiver\LogArchiver'
+    pim_connector.log_archiver:
+        class: 'Akeneo\Tool\Component\Connector\LogArchiver'
         arguments:
             - '@oneup_flysystem.archivist_filesystem'
         tags:
-            - { name: pim_connector.archiver }
+            - { name: kernel.event_subscriber }
