--- src/Akeneo/Tool/Component/Connector/Archiver/LogArchiver.php
@@ -1,70 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Akeneo\Tool\Component\Connector\Archiver;
-
-use Akeneo\Tool\Component\Batch\Model\JobExecution;
-use Akeneo\Tool\Component\Batch\Model\JobInstance;
-use League\Flysystem\Filesystem;
-
-class LogArchiver implements ArchiverInterface
-{
-    /** @var Filesystem */
-    private $filesystem;
-
-    public function __construct(Filesystem $filesystem)
-    {
-        $this->filesystem = $filesystem;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function archive(JobExecution $jobExecution): void
-    {
-        $logPath = $jobExecution->getLogFile();
-
-        if (is_file($logPath)) {
-            $log = fopen($logPath, 'r');
-            $this->filesystem->writeStream($logPath, $log);
-            if (is_resource($log)) {
-                fclose($log);
-            }
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function supports(JobExecution $jobExecution): bool
-    {
-        return in_array($jobExecution->getJobInstance()->getType(), [JobInstance::TYPE_EXPORT, JobInstance::TYPE_IMPORT]);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getArchives(JobExecution $jobExecution): array
-    {
-        $archives = [];
-
-        return $archives;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getArchive(JobExecution $jobExecution, $key)
-    {
-        // TODO: Implement getArchive() method.
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getName(): string
-    {
-        return 'log';
-    }
-}
